+++ src/GuzzleManager.php
@@ -112,7 +112,7 @@ class GuzzleManager implements Hermes
      */
     protected function resolve($name, array $options = [])
     {
-        $config = array_merge_recursive_unique(
+        $config = array_merge_recursive_distinct(
             $this->configurationFor($name), $options
         );
 
@@ -199,7 +199,7 @@ class GuzzleManager implements Hermes
     protected function createJsonDriver(array $config)
     {
         return new Client($this->makeClientOptions(
-            array_merge_recursive_unique([
+            array_merge_recursive_distinct([
                 'options' => [
                     'response_handler' => JsonResponse::class,
                 ],